test(clickhouse): rename pipeline_{mt,rmt}.rs to long-form engine names
