+++ crates/etl-destinations/tests/clickhouse/pipeline_replacing_merge_tree.rs
